requirements.txt
python-telegram-bot>=20.0
motor>=3.0.0
python-dotenv>=1.0.0
dnspython>=2.0.0
